% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicpv-package.R
\docType{package}
\name{dynamicpv-package}
\alias{dynamicpv}
\alias{dynamicpv-package}
\title{dynamicpv: Evaluates Present Values and Health Economic Models with Dynamic Pricing and Uptake}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'dynamicpv' is to provide a simple way to calculate (net) present values and outputs from health economic models (especially cost-effectiveness and budget impact) in discrete time that reflect dynamic pricing and dynamic uptake. Dynamic pricing is also known as life cycle pricing; dynamic uptake is also known as multiple or stacked cohorts, or dynamic disease prevalence. Shafrin (2024) \doi{10.1515/fhep-2024-0014} provides an explanation of dynamic value elements, in the context of Generalized Cost Effectiveness Analysis, and Puls (2024) \doi{10.1016/j.jval.2024.03.006} reviews challenges of incorporating such dynamic value elements. This package aims to reduce those challenges.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://MSDLLCpapers.github.io/dynamicpv/}
  \item \url{https://github.com/MSDLLCpapers/dynamicpv}
  \item Report bugs at \url{https://github.com/MSDLLCpapers/dynamicpv/issues}
}

}
\author{
\strong{Maintainer}: Dominic Muston \email{dominic.muston@msd.com} (\href{https://orcid.org/0000-0003-4876-7940}{ORCID})

Other contributors:
\itemize{
  \item John Blischak (\href{https://orcid.org/0000-0003-2634-9879}{ORCID}) [contributor]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder, funder]
}

}
\keyword{internal}
